Taking the first steps towards building a proper tree of conditions. This cannot be done without it
